CB-14595 Upgrade CCM Datahub handler
Upgrade CCM for all Data Hubs are launched one after the other, and the results
are polled back. If any of this is failing then the whole Environment Upgrade
CCM flow is set to failed however the poller will wait for all upgrade datahub
flows to finish (or fail).

Extend StackCcmUpgradeV4Response and DistroXCcmUpgradeC1Response to
have a status flag and an error reason as well.
Added logic to handle cases when environment is not in proper state or
the data hub is already upgraded or was not launched with a CCM tunnel type.